Laz ide compatibility changes
1) Laz IDE GDB/MI code starts debugger with following format:

$DEBUG_BINARY -i mi

where $DEBUG_BINARY is gdb or lldbmi2.

So add check of command line args for '-i mi' and start in interpreter mode.

2) GDB/MI output syntax doc (https://sourceware.org/gdb/onlinedocs/gdb/GDB_002fMI-Output-Syntax.html) indicates that if a token precedes an input command then that token should be on the output.  If no token on input then should be no token on output.

Laz IDE does not include any token on inputs and lldbmi2 interprets this as cc->sequence = 0;

Since cc->sequence is added to every cdtprintf() throughout lldbmi2, it seemed unreasonable to modify all of those instances to only output cc->sequence if > 0.  so instead I have modified cdtprintf() to strip of the opening token if it is zero.