Fix strcpy overwrite
On my OSX system, the call to strcpy() that writes the string on top of itself in StringB:clear() has a collision such that some overwriting occurs before parts of the string are copied and the string is mangled.

Fix is to copy desired portion of string into a tmp buffer first and then copy back into original string.