Change how launchInfo is created in -exec-run
With the changes to allow a process to stop without stopping target, debugger, and lldbmi2, changes must be made to -exec-run so that a target can be relaunched.  Current code uses the static version of launchInfo that retains state information from the previous launch.  This state information prevents a relaunch and cannot be cleared with current API.

Fix is to create a new launchInfo value for each launch.  It needs to be initialized with the environment parameters stored in pstate->envp, and the argv and working directory stored in the static verson of launchInfo.

Also I added an error message stirng to the target.Launch() !IsValid() return check to give more info about why process didn't launch