Modify -data-evaluate-expression and add 'strup', 'formatDesc', and 'formatStruct'
-data-evaluate-expression mods

Split off from -var-evaluate-expression and format to match 'gdb -i mi'.  This command is equivalent to 'expr' in lldb and 'print' in gdb.  Format is:

     '-data-evaluate-expression expression'

where expression can be variable names or function calls.

  'expression' can also include '.' for objects, structs, records. E.g.  'data-evaluate-expression car.brakes.areFunctioning'. SBTarget.EvaluateExpression(expression) should be able to take this type of expression and evaluate as is.  However, on OSX with lldb  library that comes with XCode this causes a segfault (also happens in plain old lldb with 'expr car.brakes.areFunctioning').  So currently this code will break up the expression at the first '.', call SBValue val =
 target.EvaluateExpression(expression1), and then call val.GetValueForExpression(expression2) to get the rest of value. This is not functional for all expressions but works for the ones I am most interested in.

This is also currently set to work with Pascal where '^' is the dereference symbol instead of '*' and it is at end instead of beginning.

- strup(char *string, int len)

Change all chars of a string to Upper Case.

- formatDesc(StringB &vardescB, SBValue var);

Walk the Base types of var and then print out a tree of the fields and methods that it inherits and contains as well as their current values (if available).

- formatStruct(StringB &vardescB, SBValue var);

Print out fields of a struct/record with types and current  values