Add functionality for -data-list-register-names and new command -data-list-register-values
-data-list-register-names

Fill in the code for this command.  Query the frame for CPU registers and return their names as an array.

-data-list-register-values.

Walk the frame register listing again and return the register index value and contents values in an array.  The index value is the index into the name array returned by -data-list-register-names associated with the specific register.