Add -data-read-memory command
-data-read-memory returns formatted contents of memory.  Format for call is:

-data-read-memory decimal-address format-type data-size num-rows num-columns

where format-type is 'x' for hex, 'd' for decimal, etc following printf convention (currently only x is supported).  data-size indicates size of access and display of data e.g. 1 = uint8, 2 = uint16, 4 = uint32, and 8 = uint64.  num-rows and num-columns indicate the number of rows and columns of data to query and display.